rotate function: complementing error treatment and tests
