+++ tests/unit_tests/visualization/test_visualize.py
@@ -7,8 +7,6 @@ import astropy.units as u
 import matplotlib.patches as mpatches
 import matplotlib.pyplot as plt
 import numpy as np
-import pytest
-from astropy.coordinates.errors import UnitsError
 
 import simtools.util.general as gen
 from simtools.visualization import visualize
@@ -109,34 +107,6 @@ def test_get_telescope_patch(corsika_telescope_data_dict):
 
         else:
             assert isinstance(patch, mpatches.Rectangle)
-
-
-def test_rotate_telescope_position():
-    x = np.array([-10.0, -10.0, 10.0, 10.0])
-    y = np.array([-10.0, 10.0, -10.0, 10.0])
-    angle_deg = 30 * u.deg
-    x_rot_manual = np.array([-13.7, -3.7, 3.7, 13.7])
-    y_rot_manual = np.array([-3.7, 13.7, -13.7, 3.7])
-
-    def check_results(x_to_test, y_to_test, x_right, y_right):
-        x_rot, y_rot = gen.rotate(angle_deg, x_to_test, y_to_test)
-        x_rot, y_rot = np.around(x_rot, 1), np.around(y_rot, 1)
-        for element, _ in enumerate(x):
-            assert x_right[element] == x_rot[element]
-            assert y_right[element] == y_rot[element]
-
-    check_results(x, y, x_rot_manual, y_rot_manual)
-
-    x_new_array, y_new_array = x * u.m, y * u.m
-    x_rot_new_array, y_rot_new_array = x_rot_manual * u.m, y_rot_manual * u.m
-    check_results(x_new_array, y_new_array, x_rot_new_array, y_rot_new_array)
-
-    with pytest.raises(TypeError):
-        gen.rotate(angle_deg, x, y[0])
-    with pytest.raises(RuntimeError):
-        gen.rotate(angle_deg, x[:-1], y)
-    with pytest.raises(UnitsError):
-        gen.rotate(angle_deg, x_new_array.to(u.cm), y_new_array)
 
 
 def test_plot_array(telescope_test_file, layout_array_north_instance):
